A bunch of improvements and fixes for tree widgets (ObserverWidget in tree view mode):
[+] ObserverWidget/TreeWidget now has a new refresh mode which makes
refreshes much smoother than previously. The new mode (which is also the
default) overlays a progress indicator on top of the current tree and
smoothly refreshes the tree, instead of hiding the tree and showing a
progress indicator instead of the tree. The old behaviour is still
around to use if you prefer to do so. See ObseverWidget::RefreshMode for
more details.
[#] ObserverWidget will no longer register all its actions by default,
only when ActionHints are specified on the displayed observer and the
ObserverWidget is set to use its observer hints, instead of custom
hints.
[#] A few optimizations and possible crash situations fixed in
ObserverTreeModel and ObserverTreeModelBuilder.
[*] Fixed ObserverWidget not respecting the
ObserverHints::ItemViewColumnFlags of new selections in TreeView mode.